renamed contacts and groups
