+++ google_contacts_api.gemspec
@@ -29,10 +29,10 @@ Gem::Specification.new do |s|
     "lib/google_contacts_api/api.rb",
     "lib/google_contacts_api/contact.rb",
     "lib/google_contacts_api/contact_set.rb",
-    "lib/google_contacts_api/contacts.rb",
+    "lib/google_contacts_api/contacts_accessor.rb",
     "lib/google_contacts_api/group.rb",
     "lib/google_contacts_api/group_set.rb",
-    "lib/google_contacts_api/groups.rb",
+    "lib/google_contacts_api/groups_accessor.rb",
     "lib/google_contacts_api/result.rb",
     "lib/google_contacts_api/result_set.rb",
     "lib/google_contacts_api/user.rb",
