+++ lib/google_contacts_api.rb
@@ -1,6 +1,6 @@
 require 'google_contacts_api/api'
-require 'google_contacts_api/contacts'
-require 'google_contacts_api/groups'
+require 'google_contacts_api/contacts_accessor'
+require 'google_contacts_api/groups_accessor'
 require 'google_contacts_api/user'
 require 'google_contacts_api/result_set'
 require 'google_contacts_api/group_set'
