+++ lib/google_contacts_api/user.rb
@@ -1,7 +1,7 @@
 module GoogleContactsApi
   class User
-    include GoogleContactsApi::Contacts
-    include GoogleContactsApi::Groups
+    include GoogleContactsApi::ContactsAccessor
+    include GoogleContactsApi::GroupsAccessor
 
     attr_reader :api
     def initialize(oauth)
@@ -14,6 +14,31 @@ module GoogleContactsApi
       @contacts ||= get_contacts(params)
     end
 
+    # def get_contacts(params = {})
+    #   puts "getting contacts"
+    #   # TODO: Should return empty ContactSet (haven't implemented one yet)
+    #   return [] unless @api
+    #   params = params.with_indifferent_access
+
+    #   # compose params into a string
+    #   # See http://code.google.com/apis/contacts/docs/3.0/reference.html#Parameters
+    #   # alt, q, max-results, start-index, updated-min,
+    #   # orderby, showdeleted, requirealldeleted, sortorder, group
+    #   params["max-results"] = 100000 unless params.key?("max-results")
+    #   url = "contacts/default/full"
+    #   response = @api.get(url, params)
+
+    #   # TODO: Define some fancy exceptions
+    #   case GoogleContactsApi::Api.parse_response_code(response)
+    #   when 401; raise
+    #   when 403; raise
+    #   when 404; raise
+    #   when 400...500; raise
+    #   when 500...600; raise
+    #   end
+    #   GoogleContactsApi::ContactSet.new(response.body, @api)
+    # end
+
     # Return the contacts for this user, retrieving them again from the server.
     def contacts!(params = {})
       # contacts in this group
@@ -25,6 +50,29 @@ module GoogleContactsApi
     def groups(params = {})
       @groups ||= get_groups(params)
     end
+
+    # def get_groups(params = {})
+    #   puts "getting groups"
+    #   params = params.with_indifferent_access
+    #   # compose params into a string
+    #   # See http://code.google.com/apis/contacts/docs/3.0/reference.html#Parameters
+    #   # alt, q, max-results, start-index, updated-min,
+    #   # orderby, showdeleted, requirealldeleted, sortorder
+    #   params["max-results"] = 100000 unless params.key?("max-results")
+
+    #   url = "groups/default/full"
+    #   response = @api.get(url, params)
+
+    #   case GoogleContactsApi::Api.parse_response_code(response)
+    #   # TODO: Better handle 401, 403, 404
+    #   when 401; raise
+    #   when 403; raise
+    #   when 404; raise
+    #   when 400...500; raise
+    #   when 500...600; raise
+    #   end
+    #   GoogleContactsApi::GroupSet.new(response.body, @api)
+    # end
 
     # Return the groups for this user, retrieving them again from the server.
     def groups!(params = {})
